Simplify the index resolution API
Removed some unused function arguments.

The arguments of as_line are simplified: If "sources" is provided, it
should be a list of dicts, and we'll use self.index to find the best one,
or just include them all. If it is left out, nothing happens.

Pipenv's caller site will need to be modified to work with this change. It
should be enough to change the call to

    sources = project.sources if include_index else None
    req = new_dep.as_line(sources=sources).strip()