Fix an issue for projects with an at sign (`@`) in the path (#310)
* Fix an issue for projects with an at sign (``@``) in the path

It seems `requirementslib` is causing a crash in `pipenv` in some rare circumstances. If you define a package like this in a `Pipfile`:
```
myproject = {editable = true, path = "."}
```
and this project path contains an at sign (`@`), then requirementslib will parse the path wrong. The second clause in the added test case will fail without this fix.

Fixes #309

* fixup! Fix an issue for projects with an at sign (``@``) in the path

* Detect @ character using a regex instead

The motivation is to only care about @ that is not part of the path but rather used for git refs

* Support paths with @ and refs with slashes

* Make the pre-commit hook happy