Remove usage of vistir.misc.run
This function is a mix of backport and yaspin spinner. The backport
part isn't needed anymore, and the yaspin spinner wasn't used.
Hence, we can safely remove that for `subporcess.run` from the STL.
This allows use to remove the modules `vistir.spin`, `vistir.colorterm`,
and `vistir._winconsole.py` from pipenv.